requirements.txt
# Core
fastapi==0.115.2
uvicorn==0.31.0
pydantic==2.9.1

# Front-end (chat UI)
streamlit==1.37.0

# Data & Memory
sqlalchemy==2.0.32
faiss-cpu==1.8.0
numpy==1.26.4

# Email & Utilities
imaplib2==3.6
python-dotenv==1.0.1
python-dateutil==2.9.0

# Behavioural Intelligence & NLP
openai>=1.43.0
nltk==3.9
schedule==1.2.2

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-playwright==0.4.4

# Code Quality
black==24.8.0
mypy==1.11.2

# Google Calendar API (Phase 3)
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0

# Misc
pyyaml==6.0.2
